No spec for `Unit`, `Boolean`, `Byte` and `Short`

+++ tests/correctness/resources/miniboxing/tests/compile/mb_normalize_8.scala
@@ -10,6 +10,6 @@ class SpCls8[@miniboxed T] {
 }
 
 object SpCls8Static {
-  def foo = new SpCls8Tgt[Byte].normalized(3, 4)
+  def foo = new SpCls8Tgt[Int].normalized(3, 4)
   def bar[@miniboxed Z](z: Z) = new SpCls8Tgt[Int].normalized(3, z)
 }
